generator: Add first integration tests
This is the first commit in a series to start adding
tests to the bindings generator.
For the tests, we will use the Regress library from gobject-introspection
since it provides a broad coverage of all possible uses cases.
https://gitlab.gnome.org/GNOME/gobject-introspection/-/tree/main/tests/scanner

The Regress-1.0.gir is processed with bindinate to generate the GAPI
xml that is feed into the code generator.
This first integration test validates the code is generated correctly.